moved time helpers into lib file
